extract hash functionality into a utility object
