+++ src/main/scala/com/github/leananeuber/hasher/actors/password_cracking/PasswordCrackingWorker.scala
@@ -1,9 +1,7 @@
 package com.github.leananeuber.hasher.actors.password_cracking
 
-import java.nio.charset.StandardCharsets
-import java.security.MessageDigest
-
 import akka.actor.{Actor, ActorLogging, Props}
+import com.github.leananeuber.hasher.HashUtil
 import com.github.leananeuber.hasher.actors.Reaper
 import com.github.leananeuber.hasher.actors.password_cracking.PasswordCrackingProtocol.{CrackPasswordsCommand, PasswordsCrackedEvent}
 import com.github.leananeuber.hasher.protocols.MasterWorkerProtocol.WorkerHandling
@@ -42,26 +40,12 @@ class PasswordCrackingWorker extends Actor with ActorLogging with WorkerHandling
   }
 
   def decrypt(secrets: Map[Int, String], range: Seq[Int]): Map[Int, Int] = {
-    val rainbow = generateRainbow(range)
+    val rainbow = HashUtil.generateRainbow(range)
     log.info("Successfully generated rainbow table! Yay!")
     for {
       idHashTuple <- secrets
-      realValue <- unhash(idHashTuple._2, rainbow)
+      realValue <- HashUtil.unhash(idHashTuple._2, rainbow)
     } yield idHashTuple._1 -> realValue
   }
 
-  private def unhash(hexHash: String, rainbow: Map[String, Int]): Option[Int] = rainbow.get(hexHash)
-
-  private def generateRainbow(range: Seq[Int]): Map[String, Int] = {
-    range.map(x => hash(x) -> x).toMap
-  }
-
-  private def hash(number: Int) = {
-    val digest = MessageDigest.getInstance("SHA-256")
-    val hashedBytes = digest.digest(String.valueOf(number).getBytes(StandardCharsets.UTF_8))
-
-    hashedBytes.map( byte =>
-      ((byte & 0xff) + 0x100).toHexString.substring(1)
-    ).mkString("")
-  }
 }
